Added users, authentication and updated api errors

+++ resolvers/account.js
@@ -2,6 +2,10 @@ import { Account } from '../models/Account';
 
 const createAccount = async (_, { account }) => {
   try {
+    const existingAccount = await Account.findOne({ userId: account.userId });
+    if (existingAccount) {
+      throw new Error(`Account with userId: ${account.userId} already exists`);
+    }
     const newAccount = new Account(account);
     await newAccount.save();
     return { account: newAccount, success: true };
@@ -11,16 +15,14 @@ const createAccount = async (_, { account }) => {
 };
 
 const editAccount = async (_, { id, account }) => {
-  const currentAccount = await Account.findById(id);
-  if (!currentAccount) {
-    return {
-      success: false
-    };
-  }
-  const mergedAccount = Object.assign(currentAccount, account);
-  mergedAccount.__v = mergedAccount.__v + 1;
-
   try {
+    const currentAccount = await Account.findById(id);
+    if (!currentAccount) {
+      throw new Error(`Account with id: ${id} does not exist`);
+    }
+    const mergedAccount = Object.assign(currentAccount, account);
+    mergedAccount.__v = mergedAccount.__v + 1;
+
     const editedAccount = await Account.findOneAndUpdate({ _id: id }, mergedAccount, {
       new: true
     });
@@ -31,9 +33,7 @@ const editAccount = async (_, { id, account }) => {
         success: true
       };
     } else {
-      return {
-        success: false
-      };
+      throw new Error('Account cannot be updated');
     }
   } catch (err) {
     throw err;
@@ -43,15 +43,17 @@ const editAccount = async (_, { id, account }) => {
 const deleteAccount = async (_, { id }) => {
   try {
     const account = await Account.findById(id);
+    if (!account) {
+      throw new Error(`Account with id: ${id} does not exist`);
+    }
+
     const response = await Account.deleteOne({ _id: id });
     if (account && response.deletedCount == 1) {
       return {
         success: true
       };
     } else {
-      return {
-        success: false
-      };
+      throw new Error('Account cannot be deleted');
     }
   } catch (err) {
     throw err;
@@ -63,6 +65,7 @@ exports.resolvers = {
     accounts: async () => Account.find(),
     account: async (_, { id }) =>
       Account.findById(id)
+        .populate({ path: 'user' })
         .populate({ path: 'bills', options: { sort: { amount: 1 } } })
         .populate({ path: 'oneOffPayments', options: { sort: { amount: 1 } } })
         .populate({ path: 'notes' })
